refactor: address review comments on chunked upload test and client
- move the ACR chunked upload integration test into adapter_test.go
- drop redundant Content-Length header now that req.ContentLength is set
- fix documented go test command path (run from the src module root)
- use a unique repository name per test run to avoid collisions
- add best-effort blob cleanup via t.Cleanup
