+++ src/pkg/registry/client.go
@@ -449,7 +449,6 @@ func (c *client) PushBlobChunk(repository, digest string, blobSize int64, chunk 
 	}
 
 	req.ContentLength = end - start + 1
-	req.Header.Set("Content-Length", fmt.Sprintf("%d", end-start+1))
 	req.Header.Set("Content-Range", fmt.Sprintf("%d-%d", start, end))
 	resp, err := c.do(req)
 	if err != nil {
